requirements.txt
flask=1.1.2
flask-wtf=0.14.3
python-dotenv=0.17.0
email_validator=1.1.2